Reformat citations
In this commit, the citations have been reformatted. This is due to the
announcement that another style will be used.

The old style automatically filled in the authors name, whilst
referencing a year. e.g:
@authorxauthory2018 -> X and Y (2018)

The new referencing should be just filling in a bracket with references
to the reference. e.g:
@authorxauthory2018 -> [1]

To reformat the citations, this has been done for all the citations in
the text:

@authorxauthory2018 -> X and Y @authorxauthory2018

In conclusion, with the new citation style, this should look as
following:

X and Y @authorxauthory2018 -> X and Y [1]